Moved protected route to components
